requirements.txt
pydub
wave==0.0.2
datetime
python-dotenv==1.0.0
